+++ src/Futhark/Representation/AST/Attributes/Ranges.hs
@@ -206,8 +206,6 @@ primOpRanges (Replicate _ v) =
   [rangeOf v]
 primOpRanges (Rearrange _ v) =
   [rangeOf $ Var v]
-primOpRanges (Split _ sizeexps v) =
-  replicate (length sizeexps) $ rangeOf $ Var v
 primOpRanges (Copy se) =
   [rangeOf $ Var se]
 primOpRanges (Index v _) =